Move `AssetTracker` to its own file.
